#19: create project skeleton, init schema, rebase common code to pda-shared for future reuse
