+++ pda-auth/src/main/java/parcel/delivery/app/auth/PdaAuthApplication.java
@@ -4,9 +4,11 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 
-@SpringBootApplication
-@ConfigurationPropertiesScan
+@SpringBootApplication(scanBasePackages = {PdaAuthApplication.APP_ROOT_PACKAGE})
+@ConfigurationPropertiesScan(basePackages = {PdaAuthApplication.APP_ROOT_PACKAGE})
 public class PdaAuthApplication {
+
+    static final String APP_ROOT_PACKAGE = "parcel.delivery.app";
 
     public static void main(String[] args) {
         SpringApplication.run(PdaAuthApplication.class, args);
