+++ pda-auth/src/main/java/parcel/delivery/app/auth/controller/AuthController.java
@@ -13,8 +13,8 @@ import parcel.delivery.app.auth.api.models.request.SignInRequest;
 import parcel.delivery.app.auth.api.models.request.SignUpRequest;
 import parcel.delivery.app.auth.api.models.response.AuthData;
 import parcel.delivery.app.auth.api.models.response.SignInResponse;
-import parcel.delivery.app.auth.security.core.UserType;
 import parcel.delivery.app.auth.service.AuthenticationService;
+import parcel.delivery.app.common.security.core.UserType;
 
 @RestController
 @RequestMapping(value = "/auth")
