--- pda-auth/src/main/java/parcel/delivery/app/auth/security/core/UserRole.java
@@ -1,5 +1,0 @@
-package parcel.delivery.app.auth.security.core;
-
-public interface UserRole {
-    String getRole();
-}
