--- pda-auth/src/main/java/parcel/delivery/app/auth/security/jwt/JwtAuthData.java
@@ -1,9 +1,0 @@
-package parcel.delivery.app.auth.security.jwt;
-
-import parcel.delivery.app.auth.security.core.RolePrivilege;
-import parcel.delivery.app.auth.security.core.UserType;
-
-import java.util.Collection;
-
-public record JwtAuthData(UserType userType, Collection<RolePrivilege> privileges) {
-}
