--- pda-auth/src/main/java/parcel/delivery/app/auth/security/jwt/JwtClaimKeys.java
@@ -1,8 +1,0 @@
-package parcel.delivery.app.auth.security.jwt;
-
-import lombok.experimental.UtilityClass;
-
-@UtilityClass
-class JwtClaimKeys {
-    static final String AUTHORIZTATION_DATA = "authorization_data";
-}
