--- pda-auth/src/main/java/parcel/delivery/app/auth/security/jwt/JwtProvider.java
@@ -1,9 +1,0 @@
-package parcel.delivery.app.auth.security.jwt;
-
-public interface JwtProvider {
-    JwtToken parse(String token);
-
-    String generate(JwtToken jwtToken);
-
-    boolean validate(String token);
-}
