+++ pda-order/src/main/java/parcel/delivery/app/order/PdaOrderApp.java
@@ -1,0 +1,11 @@
+package parcel.delivery.app.order;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PdaOrderApp {
+    public static void main(String[] args) {
+        SpringApplication.run(PdaOrderApp.class, args);
+    }
+}
