+++ pda-order/src/main/java/parcel/delivery/app/order/domain/OrderStatus.java
@@ -1,0 +1,10 @@
+package parcel.delivery.app.order.domain;
+
+public enum OrderStatus {
+    INITIAL,
+    CANCELLED,
+    ACCEPTED,
+    PENDING,
+    IN_PROGRESS,
+    DELIVERED;
+}
