+++ pda-order/src/main/resources/db/changelog-master.yaml
@@ -1,0 +1,5 @@
+databaseChangeLog:
+  - include:
+      file: changelog/20230312_001_initial_schema.yml
+      relativeToChangelogFile: true
+
