+++ pda-shared/build.gradle.kts
@@ -7,9 +7,20 @@ plugins {
 }
 
 dependencies {
+    // SPRING
+    implementation(libs.spring.web)
+    implementation(libs.spring.security)
+    implementation(libs.spring.configuration.processor)
+    // JWT
+    implementation(libs.bundles.jjwt)
+    // Lombok
+    annotationProcessor(libs.lombok)
+    compileOnly(libs.lombok)
     testCompileOnly(libs.lombok)
     testAnnotationProcessor(libs.lombok)
+    // Testing
     testImplementation(libs.junit.api)
+    testImplementation(libs.spring.test)
     testRuntimeOnly(libs.junit.engine)
 }
