--- pda-shared/src/main/java/parcel/delivery/app/Person.java
@@ -1,7 +1,0 @@
-package parcel.delivery.app;
-
-record Person(String firstName, String lastName) {
-    String fullName() {
-        return firstName + " " + lastName;
-    }
-}
