+++ pda-shared/src/main/java/parcel/delivery/app/common/security/core/RolePrivilege.java
@@ -1,0 +1,12 @@
+package parcel.delivery.app.common.security.core;
+
+import org.springframework.security.core.GrantedAuthority;
+
+public enum RolePrivilege implements GrantedAuthority {
+    CREATE_COURIER_USER;
+
+    @Override
+    public String getAuthority() {
+        return name();
+    }
+}
