+++ pda-shared/src/main/java/parcel/delivery/app/common/security/core/UserRole.java
@@ -1,0 +1,5 @@
+package parcel.delivery.app.common.security.core;
+
+public interface UserRole {
+    String getRole();
+}
