+++ pda-shared/src/main/java/parcel/delivery/app/common/security/jwt/JwtClaimKeys.java
@@ -1,0 +1,8 @@
+package parcel.delivery.app.common.security.jwt;
+
+import lombok.experimental.UtilityClass;
+
+@UtilityClass
+class JwtClaimKeys {
+    static final String AUTHORIZTATION_DATA = "authorization_data";
+}
