+++ pda-shared/src/main/java/parcel/delivery/app/common/security/jwt/JwtProvider.java
@@ -1,0 +1,9 @@
+package parcel.delivery.app.common.security.jwt;
+
+public interface JwtProvider {
+    JwtToken parse(String token);
+
+    String generate(JwtToken jwtToken);
+
+    boolean validate(String token);
+}
