+++ settings.gradle.kts
@@ -2,3 +2,4 @@ rootProject.name = "parcel-delivery-app"
 
 include("pda-shared")
 include("pda-auth")
+include("pda-order")
